feat: TossResponse -> VO 분리
DB 저장되는 값 구분 필요
