+++ src/main/kotlin/snowballclass/payment/domain/model/vo/CashReceipt.kt
@@ -1,0 +1,10 @@
+package snowballclass.payment.domain.model.vo
+
+class CashReceipt(
+	val type: String,
+	val receiptKey: String,
+	val issueNumber: String,
+	val receiptUrl: String,
+	val amount: Number,
+	val taxFreeAmount: Number
+)
