+++ src/main/kotlin/snowballclass/payment/domain/model/vo/CashReceiptItem.kt
@@ -1,0 +1,18 @@
+package snowballclass.payment.domain.model.vo
+
+class CashReceiptItem(
+	val receiptKey: String,
+	val orderId: String,
+	val orderName: String,
+	val type: String,
+	val issuerNumber: String,
+	val receiptUrl: String,
+	val businessNumber: String,
+	val transactionType: String,
+	val amount: Int,
+	val taxFreeAmount: Int,
+	val issueStatus: String,
+	val failure: Failure,
+	val customerIdentityNumber: String,
+	val requestedAt: String,
+)
