+++ src/main/kotlin/snowballclass/payment/domain/model/vo/Checkout.kt
@@ -1,0 +1,5 @@
+package snowballclass.payment.domain.model.vo
+
+class Checkout(
+	val url:String,
+)
