+++ src/main/kotlin/snowballclass/payment/domain/model/vo/Discount.kt
@@ -1,0 +1,8 @@
+package snowballclass.payment.domain.model.vo
+
+import jakarta.persistence.Embeddable
+
+@Embeddable
+class Discount(
+	val amount:Int
+)
