+++ src/main/kotlin/snowballclass/payment/domain/model/vo/Easypay.kt
@@ -1,0 +1,7 @@
+package snowballclass.payment.domain.model.vo
+
+class Easypay(
+	val provider: String,
+	val amount: Number,
+	val discountAmount: Number
+)
