+++ src/main/kotlin/snowballclass/payment/domain/model/vo/Failure.kt
@@ -1,0 +1,6 @@
+package snowballclass.payment.domain.model.vo
+
+class Failure(
+	val code:String,
+	val message:String,
+)
