+++ src/main/kotlin/snowballclass/payment/domain/model/vo/GiftCertificate.kt
@@ -1,0 +1,6 @@
+package snowballclass.payment.domain.model.vo
+
+class GiftCertificate(
+	val approveNo: String,
+	val settlementStatus: String
+)
