+++ src/main/kotlin/snowballclass/payment/domain/model/vo/MobilePhone.kt
@@ -1,0 +1,7 @@
+package snowballclass.payment.domain.model.vo
+
+class MobilePhone(
+	val customerMobilePhone: String,
+	val settlementStatus: String,
+	val receiptUrl: String
+)
