+++ src/main/kotlin/snowballclass/payment/domain/model/vo/Receipt.kt
@@ -1,0 +1,5 @@
+package snowballclass.payment.domain.model.vo
+
+class Receipt(
+	val url:String,
+)
