+++ src/main/kotlin/snowballclass/payment/domain/model/vo/Transfer.kt
@@ -1,0 +1,6 @@
+package snowballclass.payment.domain.model.vo
+
+class Transfer(
+	val bankCode: String,
+	val settlementStatus: String
+)
